+++ examples/update-management/main.tf
@@ -1,0 +1,17 @@
+terraform {
+  required_version = ">= 1.0.0"
+  required_providers {
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = ">= 2.0.0"
+    }
+    time = {
+      source  = "hashicorp/time"
+      version = ">= 0.7.2"
+    }
+  }
+}
+
+provider "azurerm" {
+  features {}
+}